Split client_def.rs in client_state, client_consensus, header
