+++ modules/src/ics02_client/client_consensus.rs
@@ -1,0 +1,127 @@
+use core::marker::{Send, Sync};
+use std::convert::TryFrom;
+
+use chrono::{DateTime, Utc};
+use prost_types::Any;
+use serde::Serialize;
+use tendermint_proto::Protobuf;
+
+use crate::ics02_client::client_type::ClientType;
+use crate::ics02_client::error::{Error, Kind};
+use crate::ics07_tendermint::consensus_state;
+use crate::ics23_commitment::commitment::CommitmentRoot;
+#[cfg(any(test, feature = "mocks"))]
+use crate::mock::client_state::MockConsensusState;
+
+pub const TENDERMINT_CONSENSUS_STATE_TYPE_URL: &str =
+    "/ibc.lightclients.tendermint.v1.ConsensusState";
+
+pub const MOCK_CONSENSUS_STATE_TYPE_URL: &str = "/ibc.mock.ConsensusState";
+
+#[dyn_clonable::clonable]
+pub trait ConsensusState: Clone + std::fmt::Debug + Send + Sync {
+    /// Type of client associated with this consensus state (eg. Tendermint)
+    fn client_type(&self) -> ClientType;
+
+    /// Commitment root of the consensus state, which is used for key-value pair verification.
+    fn root(&self) -> &CommitmentRoot;
+
+    /// Performs basic validation of the consensus state
+    fn validate_basic(&self) -> Result<(), Box<dyn std::error::Error>>;
+
+    /// Wrap into an `AnyConsensusState`
+    fn wrap_any(self) -> AnyConsensusState;
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, Serialize)]
+#[serde(tag = "type")]
+pub enum AnyConsensusState {
+    Tendermint(consensus_state::ConsensusState),
+
+    #[cfg(any(test, feature = "mocks"))]
+    Mock(MockConsensusState),
+}
+
+impl AnyConsensusState {
+    pub fn timestamp(&self) -> Result<u64, Kind> {
+        match self {
+            Self::Tendermint(cs_state) => {
+                let date: DateTime<Utc> = cs_state.timestamp.into();
+                let value = date.timestamp();
+                u64::try_from(value)
+                    .map_err(|_| Kind::NegativeConsensusStateTimestamp(value.to_string()))
+            }
+
+            #[cfg(any(test, feature = "mocks"))]
+            Self::Mock(mock_state) => Ok(mock_state.timestamp()),
+        }
+    }
+
+    pub fn client_type(&self) -> ClientType {
+        match self {
+            AnyConsensusState::Tendermint(_cs) => ClientType::Tendermint,
+
+            #[cfg(any(test, feature = "mocks"))]
+            AnyConsensusState::Mock(_cs) => ClientType::Mock,
+        }
+    }
+}
+
+impl Protobuf<Any> for AnyConsensusState {}
+
+impl TryFrom<Any> for AnyConsensusState {
+    type Error = Error;
+
+    fn try_from(value: Any) -> Result<Self, Self::Error> {
+        match value.type_url.as_str() {
+            "" => Err(Kind::EmptyConsensusStateResponse.into()),
+
+            TENDERMINT_CONSENSUS_STATE_TYPE_URL => Ok(AnyConsensusState::Tendermint(
+                consensus_state::ConsensusState::decode_vec(&value.value)
+                    .map_err(|e| Kind::InvalidRawConsensusState.context(e))?,
+            )),
+
+            #[cfg(any(test, feature = "mocks"))]
+            MOCK_CONSENSUS_STATE_TYPE_URL => Ok(AnyConsensusState::Mock(
+                MockConsensusState::decode_vec(&value.value)
+                    .map_err(|e| Kind::InvalidRawConsensusState.context(e))?,
+            )),
+
+            _ => Err(Kind::UnknownConsensusStateType(value.type_url).into()),
+        }
+    }
+}
+
+impl From<AnyConsensusState> for Any {
+    fn from(value: AnyConsensusState) -> Self {
+        match value {
+            AnyConsensusState::Tendermint(value) => Any {
+                type_url: TENDERMINT_CONSENSUS_STATE_TYPE_URL.to_string(),
+                value: value.encode_vec().unwrap(),
+            },
+            #[cfg(any(test, feature = "mocks"))]
+            AnyConsensusState::Mock(value) => Any {
+                type_url: MOCK_CONSENSUS_STATE_TYPE_URL.to_string(),
+                value: value.encode_vec().unwrap(),
+            },
+        }
+    }
+}
+
+impl ConsensusState for AnyConsensusState {
+    fn client_type(&self) -> ClientType {
+        self.client_type()
+    }
+
+    fn root(&self) -> &CommitmentRoot {
+        todo!()
+    }
+
+    fn validate_basic(&self) -> Result<(), Box<dyn std::error::Error>> {
+        todo!()
+    }
+
+    fn wrap_any(self) -> AnyConsensusState {
+        self
+    }
+}
