+++ modules/src/ics02_client/mod.rs
@@ -1,6 +1,8 @@
 //! ICS 02: IBC Client implementation
 
+pub mod client_consensus;
 pub mod client_def;
+pub mod client_state;
 pub mod client_type;
 pub mod context;
 pub mod error;
@@ -9,4 +11,3 @@ pub mod handler;
 pub mod header;
 pub mod height;
 pub mod msgs;
-pub mod state;
