+++ modules/src/ics02_client/msgs/create_client.rs
@@ -10,7 +10,8 @@ use tendermint_proto::Protobuf;
 
 use ibc_proto::ibc::core::client::v1::MsgCreateClient as RawMsgCreateClient;
 
-use crate::ics02_client::client_def::{AnyClientState, AnyConsensusState};
+use crate::ics02_client::client_consensus::AnyConsensusState;
+use crate::ics02_client::client_state::AnyClientState;
 use crate::ics02_client::error;
 use crate::ics02_client::error::{Error, Kind};
 use crate::signer::Signer;
@@ -108,7 +109,7 @@ mod tests {
 
     use ibc_proto::ibc::core::client::v1::MsgCreateClient;
 
-    use crate::ics02_client::client_def::AnyConsensusState;
+    use crate::ics02_client::client_consensus::AnyConsensusState;
     use crate::ics02_client::msgs::MsgCreateAnyClient;
 
     use crate::ics07_tendermint::client_state::test_util::get_dummy_tendermint_client_state;
