+++ modules/src/ics04_channel/context.rs
@@ -2,7 +2,8 @@
 //! the interface that any host chain must implement to be able to process any `ChannelMsg`.
 //!
 
-use crate::ics02_client::client_def::{AnyClientState, AnyConsensusState};
+use crate::ics02_client::client_consensus::AnyConsensusState;
+use crate::ics02_client::client_state::AnyClientState;
 use crate::ics03_connection::connection::ConnectionEnd;
 use crate::ics04_channel::channel::ChannelEnd;
 use crate::ics04_channel::error::Error;
