+++ modules/src/ics07_tendermint/client_def.rs
@@ -1,4 +1,6 @@
-use crate::ics02_client::client_def::{AnyClientState, AnyConsensusState, ClientDef};
+use crate::ics02_client::client_consensus::AnyConsensusState;
+use crate::ics02_client::client_def::ClientDef;
+use crate::ics02_client::client_state::AnyClientState;
 use crate::ics02_client::header::Header as ICS2Header;
 use crate::ics03_connection::connection::ConnectionEnd;
 use crate::ics04_channel::channel::ChannelEnd;
