+++ modules/src/mock/header.rs
@@ -5,11 +5,10 @@ use tendermint_proto::Protobuf;
 
 use ibc_proto::ibc::mock::Header as RawMockHeader;
 
-use crate::ics02_client::client_def::AnyConsensusState;
-use crate::ics02_client::client_def::AnyHeader;
-
+use crate::ics02_client::client_consensus::AnyConsensusState;
 use crate::ics02_client::client_type::ClientType;
 use crate::ics02_client::error::{self, Error};
+use crate::ics02_client::header::AnyHeader;
 use crate::ics02_client::header::Header;
 use crate::mock::client_state::MockConsensusState;
 use crate::Height;
