+++ relayer/src/foreign_client.rs
@@ -5,11 +5,11 @@ use thiserror::Error;
 use tracing::{error, info};
 
 use ibc::events::IbcEvent;
+use ibc::ics02_client::client_consensus::ConsensusState;
+use ibc::ics02_client::client_state::ClientState;
 use ibc::ics02_client::header::Header;
 use ibc::ics02_client::msgs::create_client::MsgCreateAnyClient;
 use ibc::ics02_client::msgs::update_client::MsgUpdateAnyClient;
-use ibc::ics02_client::state::ClientState;
-use ibc::ics02_client::state::ConsensusState;
 use ibc::ics24_host::identifier::{ChainId, ClientId};
 use ibc::tx_msg::Msg;
 use ibc::Height;
